remove static kernels;
