+++ clients/tests/accuracy_test_powX.cpp
@@ -44,7 +44,7 @@ protected:
 #define POW2_RANGE 2, 4, 8, 16, 32, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768, 65536, 131072, 262144, 1048576, 2097152, 4194304, 8388608, 16777216, 33554432
 #define POW3_RANGE 3, 9, 27, 81, 243, 729, 2187
 #define POW5_RANGE 5, 25, 125, 625, 3125
-#define MIX_RANGE 6, 10, 12, 15, 20, 30, 120, 150, 225, 240, 300, 486, 600, 900, 1250, 1500, 1875, 2160, 2187, 2250, 2500, 3000
+#define MIX_RANGE 6, 10, 12, 15, 20, 30, 120, 150, 225, 240, 300, 486, 600, 900, 1250, 1500, 1875, 2160, 2187, 2250, 2500, 3000, 4000
 
 size_t pow2_range[] = { POW2_RANGE };
 size_t pow3_range[] = { POW3_RANGE };
