+++ library/src/device/transpose.cpp
@@ -139,3 +139,67 @@ void rocfft_internal_transpose_var2(const void *data_p, void *back_p)
 
 }
 
+
+
+/* ============================================================================================ */
+
+#define TRANSPOSE_CALL(NUM_Y,DRN,TWL,TTD)    \
+hipLaunchKernel(HIP_KERNEL_NAME( transpose_var1<float2, DRN,TWL,TTD> ),\
+                    dim3(data->gridParam.b_x, data->gridParam.b_y), dim3(data->gridParam.tpb_x, data->gridParam.tpb_x), 0, 0,\
+                    (float2 *)data->node->twiddles_large, (float2 *)data->bufIn[0], (float2 *)data->bufOut[0],\
+                    NUM_Y, data->node->inStride[1], data->node->outStride[1], data->node->iDist, data->node->oDist)
+
+void rocfft_internal_transpose_var1_sp(const void *data_p, void *back_p)
+{
+    DeviceCallIn *data = (DeviceCallIn *)data_p;
+    if(data->node->transTileDir == TTD_IP_HOR)
+    {
+        if(data->node->large1D == 0)
+        {
+            if(data->node->direction == -1)
+            TRANSPOSE_CALL((data->node->length[1]/64), -1, 0, TTD_IP_HOR);
+            else
+            TRANSPOSE_CALL((data->node->length[1]/64),  1, 0, TTD_IP_HOR);
+        }
+        else if(data->node->large1D <= 16777216)
+        {
+            if(data->node->direction == -1)
+            TRANSPOSE_CALL((data->node->length[1]/64), -1, 3, TTD_IP_HOR);
+            else
+            TRANSPOSE_CALL((data->node->length[1]/64),  1, 3, TTD_IP_HOR);
+        }
+        else
+        {
+            if(data->node->direction == -1)
+            TRANSPOSE_CALL((data->node->length[1]/64), -1, 4, TTD_IP_HOR);
+            else
+            TRANSPOSE_CALL((data->node->length[1]/64),  1, 4, TTD_IP_HOR);
+        }
+    }
+    else
+    {
+        if(data->node->large1D == 0)
+        {
+            if(data->node->direction == -1)
+            TRANSPOSE_CALL((data->node->length[0]/64), -1, 0, TTD_IP_VER);
+            else
+            TRANSPOSE_CALL((data->node->length[0]/64),  1, 0, TTD_IP_VER);
+        }
+        else if(data->node->large1D <= 16777216)
+        {
+            if(data->node->direction == -1)
+            TRANSPOSE_CALL((data->node->length[0]/64), -1, 3, TTD_IP_VER);
+            else
+            TRANSPOSE_CALL((data->node->length[0]/64),  1, 3, TTD_IP_VER);
+        }
+        else
+        {
+            if(data->node->direction == -1)
+            TRANSPOSE_CALL((data->node->length[0]/64), -1, 4, TTD_IP_VER);
+            else
+            TRANSPOSE_CALL((data->node->length[0]/64),  1, 4, TTD_IP_VER);
+        }
+    }
+}
+
+
